requirements.txt
opencv-python==4.1.1.26
django-multiselectfield==0.1.9
Pillow==6.2.0